requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-simple-captcha
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv